Fail on document issue if the invoice series does not exist
Previsouly, if a series did not exist when we wanted to issue a document
with it, we just created it on the spot.

Now, to assure that any future series are created with all legal
requirements, we will fail with a specific error if a user requests to
issue a document with a unknown series.

Start adjusting tests for this change in business logic.