Updated the contexts and taxes, and removed most of the references to spain
There is still some work todo in removing references and logic from the spain module.
Also need the review the types of invoices types, the taxes percentages and other things that were not yet reviewed after copying the spain module
The next step is to finish reviewing the tests, then the same will be made to the whole logic of the actual classes.