Fixed incorrect class reference and updated tin values to remove hyphen
The main error was that the taxes were being created but the validation failed as it cannot have more then one tax with the same code.
Now, the ADFinancialValidator validates the TIN without the hyphens as given so, the tests values were updated to remove this and adjusted according to the new validation rules.