Fixes usages of toString() and Optional.of(...) in external id logic
Replace toString with getIdentifier to avoid possible overrides and
implementation changes associated with toString

Replace Optional.of() with Optional.ofNullable() to avoid potential
problems with cases where we don't have an external id. Associated with
this, improve ExternalID implementation to return null when a null value
is passed to fromValue(...).