Add inheritance annotation to all core entities that are extended in
country modules

This annotation was previously being set in JPABaseEntity but was
removed in #498. This change resulted in a change to the schema required
by the application that was only detected when a client application was
updated since the schema used during tests is generated from the code
meaning it is always correct.

Adding the annotation back to all entities that are extended in the
various submodules resolves this problem since hibernate once again
knows how the schema of the application should be assembled